Fix cross-file unit test pollution; drop assistant greet wiring suite
- Extract pure greet rules into assistantGreeting.ts so the decision test
  no longer drags in the full useAssistantChat dependency graph.
- Drop tests/test-assistant-greet-wiring.test.tsx: rendering useAssistantChat
  leaks IndexedDB/store state across suites and broke the Books suites in CI;
  the greet rules are covered by the pure decision + store tests.
- Pin Radix's layout-effect shim to a no-op in the boot screen suite so it
  behaves the same whether or not an earlier suite registered a DOM.
- Only unregister happy-dom in books-font-preview when this suite created it.
- Settle/reset files-store persist writes before wiping the ryOS database in
  the VFS metadata test, and assert only migration-owned fields.
- Update AGENTS.md: the unit aggregate is green now; document how to bisect
  order-dependent pollution.
