+++ AGENTS.md
@@ -215,4 +215,4 @@ These notes are specific to the Cursor Cloud Agent VM (where dependencies are al
 
 - **Secrets are pre-injected as environment variables** — there is **no `.env.local` file** on the cloud VM. Redis (Upstash REST: `REDIS_KV_REST_API_URL`/`REDIS_KV_REST_API_TOKEN`), Pusher, and AI keys (OpenAI/Anthropic/Google) are already present, so `bun run dev` works with full functionality out of the box. The README's `.env.local` instructions are for local maintainer machines only.
 - **AI chat is rate-limited, which can look like a broken AI.** `/api/chat` (the Ryo assistant in the Chats app) allows only **3 messages/day for anonymous users (per IP)** and **15 per 5h for authenticated users (per username)** (see `api/_utils/_rate-limit.ts`). The shared anonymous per-IP budget is quickly exhausted by the API test suite, so if Ryo doesn't reply while logged out you're almost certainly rate-limited (HTTP 429), not missing keys. **Sign in (register/login) before testing AI in the UI.**
-- **`bun run test:unit` aggregate run has pre-existing failures unrelated to the environment.** Running all unit suites in one process currently yields ~21 deterministic failures from cross-file state pollution (zustand persist + happy-dom storage), reproducible across bun 1.3.5/1.3.9/1.3.14. Each affected suite **passes in isolation** (e.g. `bun test ./tests/test-app-store-foreground-state.test.ts`). Before treating a unit failure as a regression, re-run the specific suite on its own. The full API integration suite (`bun run dev:api` + `bun run test:api`) passes 325/325.
+- **`bun run test:unit` runs every suite in one Bun process, so cross-file pollution is the usual cause of aggregate-only failures.** Bun's file execution order differs between machines (it is not the CLI/alphabetical order), so a suite can pass in CI yet fail locally, or vice versa. If an aggregate failure does not reproduce in isolation (`bun test ./tests/<suite>.test.ts`), bisect the actual execution order (the `tests/<file>:` headers in the log) to find the polluting suite instead of treating it as a regression. Known traps: unregistering a happy-dom `GlobalRegistrator` another suite registered; Radix UI's layout-effect shim freezing its DOM detection at first import (`tests/test-boot-screen-accent.test.ts` pins it via `mock.module`); leaked `useFilesStore` state/debounced IndexedDB writes from earlier suites (settle + reset before `deleteDatabase`, which silently no-ops when blocked). The full API integration suite (`bun run dev:api` + `bun run test:api`) passes 325/325.
